Adds category names as prefix to node filenames
