+++ nodes/number_number_range.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 ###################################################################################
 #
-#  get_objects_data.py
+#  number_number_range.py
 #
 #  Copyright (c) 2022 Ronny Scharf-Wildenhain <ronny.scharf08@gmail.com>
 #
@@ -22,7 +22,9 @@
 #
 #
 ###################################################################################
-import FreeCAD
+import os
+import awkward as ak
+import numpy as np
 
 from fcn_conf import register_node
 from fcn_base_node import FCNNode
@@ -30,41 +32,33 @@ from fcn_locator import icon
 
 
 @register_node
-class GetObjectsData(FCNNode):
+class NumberRange(FCNNode):
 
     icon: str = icon("fcn_default.png")
-    op_title: str = "Get Objects Data"
-    op_category = "Scene"
+    op_title: str = "Number Range"
+    op_category = "Number"
     content_label_objname: str = "fcn_node_bg"
 
     def __init__(self, scene):
         super().__init__(scene=scene,
-                         inputs_init_list=[(3, "In", 1, "Object label", True, ("str", ))],
-                         outputs_init_list=[(4, "Obj", 0, 0, True, ("fc_obj", ))],
-                         width=170)
-
-    def collapse_node(self, collapse: bool = False):
-        super().collapse_node(collapse)
-
-        if collapse is True:
-            self.title = 'Obj: ' + str(self.sockets_input_data[0][0])
-        else:
-            self.title = self.default_title
+                         inputs_init_list=[(0, "Start", 1, 0, True, ('int', 'float')),
+                                           (0, "Stop", 1, 10, True, ('int', 'float')),
+                                           (0, "Step", 1, 1, False, ('int', 'float'))],
+                         outputs_init_list=[(0, "Out", 0, 0, True, ('int', 'float'))],
+                         width=150)
 
     def eval_operation(self, sockets_input_data: list) -> list:
-        self.collapse_node(self.content.isHidden())
+        # Inputs
+        start = sockets_input_data[0]
+        stop = sockets_input_data[1]
+        step = sockets_input_data[2][0]
 
-        label_list: list = sockets_input_data[0]
-        obj_list: list = []
+        # Force array broadcast
+        start, stop = ak.broadcast_arrays(start, stop)
 
-        if not (FreeCAD.ActiveDocument is None):
-            for label in label_list:
-                obj = FreeCAD.ActiveDocument.getObjectsByLabel(label)
-                if len(obj) == 1:
-                    obj_list.append(obj[0])
-                else:
-                    raise ValueError('Unknown object')
-        else:
-            raise Exception('No active document')
+        res = []
+        for idx, _start in enumerate(ak.flatten(start, axis=None)):
+            _stop = ak.flatten(stop, axis=None)[idx]
+            res.append(np.arange(_start, _stop, step).tolist())
 
-        return [obj_list]
+        return [res]
